refactor: 코드 가독성을 고려해 Goal API를 클래스별로 분리
