+++ src/main/java/com/groom/orbit/goal/membergoal/controller/MemberGoalQueryController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.*;
 import com.groom.orbit.common.annotation.AuthMember;
 import com.groom.orbit.common.dto.ResponseDto;
 import com.groom.orbit.goal.goal.application.dto.response.GetMemberGoalResponseDto;
-import com.groom.orbit.goal.membergoal.application.MemberGoalService;
+import com.groom.orbit.goal.membergoal.application.MemberGoalQueryService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -16,18 +16,18 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/api/goal")
 public class MemberGoalQueryController {
 
-  private final MemberGoalService memberGoalService;
+  private final MemberGoalQueryService memberGoalQueryService;
 
   @GetMapping
   public ResponseDto<List<GetMemberGoalResponseDto>> getGoals(
       @AuthMember Long memberId,
       @RequestParam(value = "is_complete", required = false) Boolean isComplete) {
-    return ResponseDto.ok(memberGoalService.findGoals(memberId, isComplete));
+    return ResponseDto.ok(memberGoalQueryService.findGoals(memberId, isComplete));
   }
 
   @GetMapping("/{member_goal_id}")
   public ResponseDto<GetMemberGoalResponseDto> getGoal(
       @PathVariable("member_goal_id") Long memberGoalId) {
-    return ResponseDto.ok(memberGoalService.findGoal(memberGoalId));
+    return ResponseDto.ok(memberGoalQueryService.findGoal(memberGoalId));
   }
 }
